+++ src/components/Main/PersonalFeedList.tsx
@@ -1,8 +1,8 @@
 import { Box, CircularProgress, Fade, Stack } from '@mui/material';
 import { useMemo } from 'react';
 
-import { Title } from '@/components/common';
 import { Feed } from '@/components/common/Feed';
+import { Title } from '@/components/common/Title';
 import useIntersect from '@/hooks/common/useIntersect';
 import { useGetPersonalTravelogues } from '@/hooks/query/main';
 import { flexCenterStyle } from '@/styles/commonStyle';
