+++ src/components/common/SwipeSlider.tsx
@@ -4,8 +4,8 @@ import 'slick-carousel/slick/slick-theme.css';
 import { Stack, SxProps, Theme } from '@mui/material';
 import Slider from 'react-slick';
 
-import { Title } from '@/components/common';
 import { Feed } from '@/components/common/Feed';
+import { Title } from '@/components/common/Title';
 import { TravelogueFeedType } from '@/types/main';
 
 interface SwipeSliderProps {
