+++ src/components/detail/TravelogueInfo.tsx
@@ -1,7 +1,7 @@
 import { Visibility as VisibilityIcon } from '@mui/icons-material';
 import { Stack, Typography } from '@mui/material';
 
-import UserProfile from '@/components/common/UserProfile/UserProfile';
+import { UserProfile } from '@/components/common/UserProfile';
 import { TravelogueDetailType } from '@/types/detail';
 
 interface PostProfileProps {
