+++ src/components/travelogue/Travelogue/PostBasic.tsx
@@ -3,12 +3,13 @@ import dayjs from 'dayjs';
 import { MouseEvent, useEffect, useState } from 'react';
 import { Control, Controller } from 'react-hook-form';
 
-import { FileInput, SubTitle, Title } from '@/components/common';
-import { CountrySelect } from '@/components/common';
+import { FileInput } from '@/components/common';
+import { SubTitle, Title } from '@/components/common/Title';
 import useTravelogueForm from '@/hooks/travelogue/useTravelogueForm';
 import { inputStyle } from '@/styles/commonStyle';
 import { TravelogueFormType, TravelogueResponseType } from '@/types/travelogue';
 
+import { CountrySelect } from '.';
 import { ComplexButton, DatePicker } from '.';
 
 interface PostBasicProps {
