+++ src/pages/_app.tsx
@@ -3,10 +3,8 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import type { AppProps } from 'next/app';
 import { RecoilRoot } from 'recoil';
 
-import { SigninLeadModal } from '@/components/common';
+import { GNB, Header, LoginConfirmModal } from '@/components/common';
 import { AutoComplete } from '@/components/common/AutoComplete';
-import { GNB } from '@/components/common/GNB';
-import { Header } from '@/components/common/Header';
 import GlobalStyle from '@/styles/GlobalStyle';
 import MobileLayout from '@/styles/MobileLayout';
 import { theme } from '@/styles/MuiTheme';
@@ -33,7 +31,7 @@ export default function App({ Component, pageProps }: AppProps) {
             </AutoComplete>
             <Component {...pageProps} />
             <GNB />
-            <SigninLeadModal />
+            <LoginConfirmModal />
           </MobileLayout>
         </RecoilRoot>
       </ThemeProvider>
