+++ src/pages/auth/register.tsx
@@ -3,8 +3,8 @@ import { useRouter } from 'next/router';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 
+import { AuthHeader } from '@/components/auth';
 import { Register, VerifyByEmail } from '@/components/auth/Register';
-import { AuthHeader } from '@/components/common';
 import { Stepper } from '@/components/common/Stepper';
 import { usePostUserRegister } from '@/hooks/query/auth';
 import { UserRegister } from '@/types/auth';
